requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.5
    # via discord-py
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
black==23.9.1
    # via -r requirements.in
build==1.0.3
    # via pip-tools
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.2.0
    # via aiohttp
click==8.1.7
    # via
    #   black
    #   pip-tools
discord-py==2.3.2
    # via -r requirements.in
distlib==0.3.7
    # via virtualenv
filelock==3.12.3
    # via virtualenv
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
identify==2.5.28
    # via pre-commit
idna==3.4
    # via yarl
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via black
nodeenv==1.8.0
    # via pre-commit
packaging==23.1
    # via
    #   black
    #   build
pathspec==0.11.2
    # via black
pip-tools==7.3.0
    # via -r requirements.in
platformdirs==3.10.0
    # via
    #   black
    #   virtualenv
pre-commit==3.4.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
python-dotenv==1.0.0
    # via -r requirements.in
pyyaml==6.0.1
    # via pre-commit
ruff==0.0.288
    # via -r requirements.in
tomli==2.0.1
    # via
    #   black
    #   build
    #   pip-tools
    #   pyproject-hooks
typing-extensions==4.7.1
    # via
    #   black
    #   filelock
virtualenv==20.24.5
    # via pre-commit
wheel==0.41.2
    # via pip-tools
yarl==1.9.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
